requirements.txt
APScheduler==3.6.3
astroid==2.4.2
autopep8==1.5.4
certifi==2020.11.8
cffi==1.14.3
cryptography==3.2.1
decorator==4.4.2
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycodestyle==2.6.0
pycparser==2.20
pylint==2.6.0
python-dotenv==0.15.0
python-telegram-bot==13.0
pytz==2020.4
six==1.15.0
toml==0.10.2
tornado==6.1
tzlocal==2.1
wrapt==1.12.1
